requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgi-redis
asgiref
asn1crypto==0.24.0
astroid==2.4.2
async-timeout==2.0.1
attrs==19.3.0
autobahn==19.6.2
Automat==0.7.0
backports.csv==1.0.7
base58check==1.0.2
billiard==3.6.0.0
celery==5.0.5
certifi==2019.6.16
cffi==1.12.3
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
coinaddr==1.0.1
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
Django==3.0.5
django-3-jet==1.0.8
django-autoslug==1.9.4
django-cors-headers==3.5.0
django-crequest==2018.5.11
django-environ==0.4.5
django-markdownx==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
drf-nested-routers==0.92.1
drf-yasg==1.17.1
email-validator==1.0.4
et-xmlfile==1.0.1
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.3
lazy-object-proxy==1.4.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
oauthlib==3.0.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
packaging==20.4
password-strength==0.0.3.post2
pep8==1.7.1
Pillow==7.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyotp==2.4.0
pyparsing==2.4.7
pysha3==1.0.2
python-dateutil==2.8.0
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2019.1
PyYAML==5.3.1
qrcode==6.1
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sentry-sdk==0.18.0
simplejson==3.17.2
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.1
Twisted==20.3.0
txaio==18.8.1
typed-ast==1.4.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.10
user-agents==2.2.0
vine==1.3.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.2.0
